# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
...
    "/pages": {
      "post": {
        "operationId": "create_pages",
        "requestBody": {
          "content": {
            "multipart/form-data": {}
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePagesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/pages/from-paths": {
      "post": {
        "summary": "Create pages by reading image files from absolute paths on the server's\nfilesystem. This is the Tauri desktop import path — the webview picker\nreturns paths, and the backend reads + decodes + hashes them in parallel\nwithout a round-trip through JS memory or a multipart upload body.",
        "description": "Web clients should keep using `POST /pages` with multipart.",
        "operationId": "create_pages_from_paths",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePagesFromPathsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePagesResponse"
                }
              }
            }
          }
        }
      }
    },
    "/pages/{id}/masks/{role}": {
      "put": {
        "summary": "Upsert the `Mask { role }` node on a page with the raw image bytes in\nthe body. Emits `Op::UpdateNode` if a mask of that role exists, else\n`Op::AddNode`. Used by the repair-brush / segment-edit flow; the\nfollow-up localized inpaint is a separate `POST /pipelines` call.",
        "operationId": "put_mask",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Page id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageId"
            }
          },
          {
            "name": "role",
            "in": "path",
            "description": "Mask role (segment|brushInpaint)",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/MaskRole"
            }
          },
          {
            "name": "pipeline",
            "in": "query",
            "description": "Optional pipeline engine to run after the mask is updated.",
            "required": false,
            "schema": {
              "type": ["string", "null"]
            }
          },
          {
            "name": "x",
            "in": "query",
            "description": "Bounding box for the pipeline run.",
            "required": false,
            "schema": {
              "type": ["number", "null"],
              "format": "float"
            }
          },
          {
            "name": "y",
            "in": "query",
            "required": false,
            "schema": {
              "type": ["number", "null"],
              "format": "float"
            }
          },
          {
            "name": "width",
            "in": "query",
            "required": false,
            "schema": {
              "type": ["number", "null"],
              "format": "float"
            }
          },
          {
            "name": "height",
            "in": "query",
            "required": false,
            "schema": {
              "type": ["number", "null"],
              "format": "float"
            }
          }
        ],
        "requestBody": {
          "content": {
            "image/png": {}
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PutMaskResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/pages/{page_id}/reorder-text-nodes": {
      "post": {
        "operationId": "reorder_text_nodes",
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "description": "Page id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PageId"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadingOrder"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/pipelines": {
      "post": {
        "operationId": "start_pipeline",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPipelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartPipelineResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlobRef": {
        "type": "string",
        "description": "Hex-encoded blake3 hash of an immutable blob."
      },
...
      "CreatePagesFromPathsRequest": {
        "type": "object",
        "required": ["paths"],
        "properties": {
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "replace": {
            "type": "boolean"
          }
        }
      },
      "CreatePagesResponse": {
        "type": "object",
        "required": ["pages"],
        "properties": {
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PageId"
            }
          }
        }
      },
...
      "MaskRole": {
        "type": "string",
        "enum": ["brushInpaint", "segment", "bubble"]
      },
...
      "NodeId": {
        "type": "string",
        "format": "uuid"
      },
...
      "PageId": {
        "type": "string",
        "format": "uuid"
      },
...
      "PutMaskResponse": {
        "type": "object",
        "required": ["node", "blob"],
        "properties": {
          "blob": {
            "$ref": "#/components/schemas/BlobRef"
          },
          "node": {
            "$ref": "#/components/schemas/NodeId"
          }
        }
      },
      "ReadingOrder": {
        "type": "string",
        "enum": ["rtl", "ltr", "custom"]
      },
      "Region": {
        "type": "object",
        "required": ["x", "y", "width", "height"],
        "properties": {
          "height": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "width": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "x": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "y": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "StartPipelineRequest": {
        "type": "object",
        "required": ["steps"],
        "properties": {
          "defaultFont": {
            "type": ["string", "null"]
          },
          "pages": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/PageId"
            },
            "description": "`None` → whole project, `Some(pages)` → just those pages."
          },
          "readingOrder": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReadingOrder"
              }
            ]
          },
          "region": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Region",
                "description": "Optional bounding-box hint for inpainter engines (repair-brush)."
              }
            ]
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Engine ids (`inventory::submit!` ids) to run in order."
          },
          "systemPrompt": {
            "type": ["string", "null"]
          },
          "targetLanguage": {
            "type": ["string", "null"]
          },
          "textNodeIds": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/NodeId"
            },
            "description": "Optional text-node ids for engines that can operate on individual blocks."
          }
        }
      },
      "StartPipelineResponse": {
        "type": "object",
        "required": ["operationId"],
        "properties": {
          "operationId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}